@@ -50,62 +49,59 @@
   },
   "homepage": "https://angular-maps.com",
   "peerDependencies": {
-    "@angular/common": "^4.0.0 || ^5.0.0",
-    "@angular/core": "^4.0.0 || ^5.0.0"
+    "@angular/common": "^5.0.0 || ^6.0.0",
+    "@angular/core": "^5.0.0 || ^6.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/platform-server": "^4.0.0",
-    "@angular/tsc-wrapped": "4.0.0",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/platform-server": "^6.0.0",
+    "@angular/tsc-wrapped": "4.4.6",
     "@compodoc/compodoc": "^1.0.0-beta.9",
-    "@ngtools/webpack": "1.3.0",
-    "@types/jasmine": "2.5.46",
+    "@ngtools/webpack": "6.0.0",
+    "@types/jest": "^23.0.0",
     "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "3.1.2",
-    "clang-format": "1.0.46",
-    "codelyzer": "^3.2.2",
+    "awesome-typescript-loader": "^5.0.0",
+    "clang-format": "1.2.3",
+    "codelyzer": "^4.3.0",
     "conventional-changelog": "^1.1.0",
     "copy-webpack-plugin": "^4.0.1",
-    "copyfiles": "^1.0.0",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
+    "copyfiles": "^2.0.0",
     "glob": "^7.1.2",
-    "html-webpack-plugin": "^2.8.1",
-    "istanbul-instrumenter-loader": "^2.0.0",
-    "jasmine-core": "2.5.0",
+    "html-webpack-plugin": "^3.2.0",
+    "istanbul-instrumenter-loader": "^3.0.1",
+    "jasmine-core": "3.1.0",
+    "jest": "^23.1.0",
+    "jest-zone-patch": "^0.0.8",
     "js-marker-clusterer": "^1.0.0",
-    "karma": "^1.3.0",
-    "karma-chrome-launcher": "2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-mocha-reporter": "^2.2.0",
-    "karma-remap-istanbul": "^0.5.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.8.0",
+    "karma-chrome-launcher": "2.2.0",
     "mkdirp": "^0.5.1",
     "ncp": "^2.0.0",
     "reflect-metadata": "^0.1.8",
-    "remap-istanbul": "^0.9.1",
-    "rimraf": "2.6.1",
-    "rollup": "0.41.6",
-    "rxjs": "^5.2.0",
+    "remap-istanbul": "^0.11.1",
+    "rimraf": "2.6.2",
+    "rollup": "0.58.2",
+    "rxjs": "^6.1.0",
+    "snazzy-info-window": "^1.1.1",
     "ts-helpers": "^1.1.2",
+    "ts-jest": "^22.4.6",
+    "ts-loader": "^4.3.0",
     "tslint": "^5.8.0",
     "tslint-loader": "^3.4.3",
-    "typescript": "~2.5.0",
-    "webpack": "2.3.3",
-    "zone.js": "^0.8.4"
+    "typescript": "~2.7.2",
+    "webpack": "4.8.0",
+    "zone.js": "^0.8.26"
   },
   "jspm": {
     "jspmNodeConversion": false,
     "dependencies": {
-      "@angular/common": "^4.0.0 || ^5.0.0",
-      "@angular/core": "^4.0.0 || ^5.0.0"
+      "@angular/common": "^5.0.0 || ^6.0.0",
+      "@angular/core": "^5.0.0 || ^6.0.0"
     }
   }
 }
