@@ -1,7 +1,7 @@
 import {Directive, EventEmitter, OnChanges, OnDestroy, SimpleChange,
   AfterContentInit, ContentChildren, QueryList, Input, Output
 } from '@angular/core';
-import {Subscription} from 'rxjs/Subscription';
+import {Subscription} from 'rxjs';
 
 import {MouseEvent} from '../map-types';
 import * as mapTypes from '../services/google-maps-types';
@@ -116,6 +116,11 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
   @Output() markerClick: EventEmitter<void> = new EventEmitter<void>();
 
   /**
+   * This event is fired when the user rightclicks on the marker.
+   */
+  @Output() markerRightClick: EventEmitter<void> = new EventEmitter<void>();
+
+  /**
    * This event is fired when the user stops dragging the marker.
    */
   @Output() dragEnd: EventEmitter<MouseEvent> = new EventEmitter<MouseEvent>();
@@ -158,6 +163,12 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
 
   /** @internal */
   ngOnChanges(changes: {[key: string]: SimpleChange}) {
+    if (typeof this.latitude === 'string') {
+      this.latitude = Number(this.latitude);
+    }
+    if (typeof this.longitude === 'string') {
+      this.longitude = Number(this.longitude);
+    }
     if (typeof this.latitude !== 'number' || typeof this.longitude !== 'number') {
       return;
     }
@@ -207,6 +218,11 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
       this.markerClick.emit(null);
     });
     this._observableSubscriptions.push(cs);
+
+    const rc = this._markerManager.createEventObservable('rightclick', this).subscribe(() => {
+      this.markerRightClick.emit(null);
+    });
+    this._observableSubscriptions.push(rc);
 
     const ds =
         this._markerManager.createEventObservable<mapTypes.MouseEvent>('dragend', this)
