@@ -13,7 +13,9 @@ describe('MarkerManager', () => {
         {provide: NgZone, useFactory: () => new NgZone({enableLongStackTrace: true})},
         MarkerManager, {
           provide: GoogleMapsAPIWrapper,
-          useValue: jasmine.createSpyObj('GoogleMapsAPIWrapper', ['createMarker'])
+          useValue: {
+            createMarker: jest.fn()
+          }
         }
       ]
     });
@@ -55,8 +57,10 @@ describe('MarkerManager', () => {
              newMarker.longitude = 22.3;
              newMarker.label = 'A';
 
-             const markerInstance: Marker = jasmine.createSpyObj('Marker', ['setMap']);
-             (<any>apiWrapper.createMarker).and.returnValue(Promise.resolve(markerInstance));
+             const markerInstance: any = {
+              setMap: jest.fn()
+             };
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
 
              markerManager.addMarker(newMarker);
              markerManager.deleteMarker(newMarker).then(
@@ -74,8 +78,11 @@ describe('MarkerManager', () => {
              newMarker.longitude = 22.3;
              newMarker.label = 'A';
 
-             const markerInstance: Marker = jasmine.createSpyObj('Marker', ['setMap', 'setIcon']);
-             (<any>apiWrapper.createMarker).and.returnValue(Promise.resolve(markerInstance));
+             const markerInstance: any = {
+              setMap: jest.fn(),
+              setIcon: jest.fn()
+             };
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
 
              markerManager.addMarker(newMarker);
              expect(apiWrapper.createMarker).toHaveBeenCalledWith({
@@ -107,9 +114,13 @@ describe('MarkerManager', () => {
              newMarker.longitude = 22.3;
              newMarker.label = 'A';
 
-             const markerInstance: Marker =
-                 jasmine.createSpyObj('Marker', ['setMap', 'setOpacity']);
-             (<any>apiWrapper.createMarker).and.returnValue(Promise.resolve(markerInstance));
+             const markerInstance: any = {
+              setMap: jest.fn(),
+              setOpacity: jest.fn()
+             };
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
+
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
 
              markerManager.addMarker(newMarker);
              expect(apiWrapper.createMarker).toHaveBeenCalledWith({
@@ -142,9 +153,11 @@ describe('MarkerManager', () => {
              newMarker.label = 'A';
              newMarker.visible = false;
 
-             const markerInstance: Marker =
-                 jasmine.createSpyObj('Marker', ['setMap', 'setVisible']);
-             (<any>apiWrapper.createMarker).and.returnValue(Promise.resolve(markerInstance));
+             const markerInstance: any = {
+              setMap: jest.fn(),
+              setVisible: jest.fn()
+             };
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
 
              markerManager.addMarker(newMarker);
              expect(apiWrapper.createMarker).toHaveBeenCalledWith({
@@ -176,8 +189,11 @@ describe('MarkerManager', () => {
              newMarker.label = 'A';
              newMarker.visible = false;
 
-             const markerInstance: Marker = jasmine.createSpyObj('Marker', ['setMap', 'setZIndex']);
-             (<any>apiWrapper.createMarker).and.returnValue(Promise.resolve(markerInstance));
+             const markerInstance: any = {
+              setMap: jest.fn(),
+              setZIndex: jest.fn()
+             };
+             (<jest.Mock>apiWrapper.createMarker).mockReturnValue(Promise.resolve(markerInstance));
 
              markerManager.addMarker(newMarker);
              expect(apiWrapper.createMarker).toHaveBeenCalledWith({
