@@ -1,22 +1,24 @@
 export default {
-  entry: 'dist/core/index.js',
-  dest: 'dist/core/core.umd.js',
-  format: 'umd',
-  moduleName: 'ngmaps.core',
-  sourceMap: true,
-  globals: {
-    '@angular/core': 'ng.core',
-    '@angular/common': 'ng.common',
-    '@angular/compiler': 'ng.compiler',
-    '@angular/platform-browser': 'ng.platformBrowser',
-    '@angular/platform-browser-dynamic': 'ng.platformBrowserDynamic',
-    'js-marker-clusterer': 'MarkerClusterer',
-    'rxjs/Subject': 'Rx',
-    'rxjs/observable/PromiseObservable': 'Rx',
-    'rxjs/operator/toPromise': 'Rx.Observable.prototype',
-    'rxjs/Observable': 'Rx',
-    'rxjs/Rx': 'Rx'
+  input: 'dist/packages/core/index.js',
+  output: {
+    file: 'dist/packages/core/core.umd.js',
+    name: 'ngmaps.core',
+    globals: {
+      '@angular/core': 'ng.core',
+      '@angular/common': 'ng.common',
+      '@angular/compiler': 'ng.compiler',
+      '@angular/platform-browser': 'ng.platformBrowser',
+      '@angular/platform-browser-dynamic': 'ng.platformBrowserDynamic',
+      'js-marker-clusterer': 'MarkerClusterer',
+      'rxjs/Subject': 'Rx',
+      'rxjs/observable/PromiseObservable': 'Rx',
+      'rxjs/operator/toPromise': 'Rx.Observable.prototype',
+      'rxjs/Observable': 'Rx',
+      'rxjs/Rx': 'Rx'
+    },
+    sourceMap: true,
+    format: 'umd'
   },
   context: 'window',
-  external: ['rxjs', '@angular/core', 'rxjs/Observable', 'js-marker-clusterer']
+  external: ['rxjs', '@angular/core']
 }
